feat: clean dependencies on reruns (#31)
Shamelessly stolen from this Felt action:
https://github.com/felt/ultimate-elixir-ci/blob/main/.github/actions/elixir-setup/action.yml#L97

If a person reruns a GitHub action, it will cause Elixir to clean all
dependencies and re fetch them. This will trigger a rebuild in what ever
action using this composite action. This makes a nice alternative to
having to clear the cache if there is an issue.